Add files last changed on 1st November 2020
I never version-controlled this little project back when I created it, so now I'm just uploading it as a couple of files. It took exactly 7 days to create this interesting chat solution. After creating it, I chatted with a friend of mine over a college Linux server for like 5 hours and never used it again.